pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "eve"
version = "0.1.0"
description = "Eve - co-evolution of solver and optimizer agents for scientific search"
requires-python = ">=3.11,<3.15"
license = { text = "Apache-2.0" }
authors = [{ name = "Scientific Computing and Intelligence Group" }]
dependencies = [
  "hydra-core>=1.3",
  "numpy>=2.2",
  "omegaconf>=2.3",
  "pydantic>=2.11",
  "python-dotenv>=1.2.2",
  "PyYAML>=6.0.2",
  "wandb>=0.19",
  "optree>=0.19.0",
]

[dependency-groups]
dev = [
  "pre-commit>=4.2",
  "pytest>=8.3",
  "pytest-cov>=6.0",
  "pytest-xdist>=3.6",
  "ruff>=0.11",
  "types-PyYAML>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/scaling_evolve"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101"]

[tool.pytest.ini_options]
addopts = "-q --strict-markers"
testpaths = ["tests"]
norecursedirs = ["tests/fixtures/repos"]
pythonpath = ["src"]
markers = [
  "integration: slower cross-module tests",
  "live: requires external API key or manual trigger",
]
